requirements.txt
# Core dependencies for local LLM
# Python 3.10+ required

# PyTorch (GPU support)
torch>=2.0.0

# Hugging Face ecosystem
transformers>=4.35.0
accelerate>=0.25.0
tokenizers>=0.15.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Configuration
python-dotenv>=1.0.0

# Terminal UI (Phase 4 - Enhanced CLI)
rich>=13.0.0
prompt-toolkit>=3.0.0

# Testing (for Phase 6)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional: Quantization and optimization
# bitsandbytes>=0.41.0  # Uncomment for 4-bit/8-bit quantization
# safetensors>=0.4.0    # Uncomment for safer model serialization
